requirements.txt
yfinance==0.2.32
scikit-learn==1.3.0
pandas==2.0.3
apache-airflow==2.7.3
mlflow==2.9.1
